refactor: Remove unnecessary __main__.py file
cloud-init's subcommands are not expected to be executable. This file
was added in e74d7752f17 when an external analyzer tool was integrated into
cloud-init as a subcommand. Presumably this file was previously executed
as a submodule via `python3 -m analyze`. Since this is not a requirement
since integration into cloud-init, module initialization code belongs in
__init__.py
